refactor: move health target contracts to output crate
