+++ crates/output/src/lib.rs
@@ -17,6 +17,7 @@ mod codeclimate;
 mod dupes;
 mod format;
 mod health;
+mod health_targets;
 mod issue_contract;
 mod report_contract;
 
@@ -40,6 +41,11 @@ pub use fallow_types::output_dead_code;
 pub use fallow_types::output_health;
 pub use format::OutputFormat;
 pub use health::{HealthOutput, HealthOutputInput, build_health_output};
+pub use health_targets::{
+    CloneSiblingEvidence, Confidence, ContributingFactor, DirectCallerEvidence,
+    DirectCallerSymbolEvidence, EffortEstimate, EvidenceFunction, RecommendationCategory,
+    RefactoringTarget, TargetEvidence, TargetThresholds,
+};
 pub use issue_contract::{
     ACTIONS_AUTO_FIXABLE_FIELD_DEFINITION, ACTIONS_FIELD_DEFINITION, CHECK_DOCS,
     CODECLIMATE_RESULT_CODES, IssueOutputContract, TsAliasMeta, check_meta, dead_code_docs_url,
